Now using class instead of PlaylistOwnerEntity interface
