+++ src/main/java/com/odeyalo/sonata/playlists/repository/r2dbc/callback/read/PlaylistOwnerAssociationAfterConvertCallback.java
@@ -1,6 +1,7 @@
 package com.odeyalo.sonata.playlists.repository.r2dbc.callback.read;
 
 import com.odeyalo.sonata.playlists.entity.PlaylistEntity;
+import com.odeyalo.sonata.playlists.entity.PlaylistOwnerEntity;
 import com.odeyalo.sonata.playlists.repository.R2dbcPlaylistOwnerRepository;
 import org.jetbrains.annotations.NotNull;
 import org.reactivestreams.Publisher;
@@ -10,7 +11,7 @@ import org.springframework.data.relational.core.sql.SqlIdentifier;
 import org.springframework.stereotype.Component;
 
 /**
- * Associate the {@link PlaylistEntity} with the {@link com.odeyalo.sonata.playlists.entity.R2dbcPlaylistOwnerEntity}
+ * Associate the {@link PlaylistEntity} with the {@link PlaylistOwnerEntity}
  */
 @Component
 public final class PlaylistOwnerAssociationAfterConvertCallback implements AfterConvertCallback<PlaylistEntity> {
