+++ src/test/java/testing/faker/PlaylistEntityFaker.java
@@ -1,9 +1,8 @@
 package testing.faker;
 
 import com.github.javafaker.Faker;
-import com.odeyalo.sonata.playlists.entity.PlaylistOwnerEntity;
 import com.odeyalo.sonata.playlists.entity.PlaylistEntity;
-import com.odeyalo.sonata.playlists.entity.R2dbcPlaylistOwnerEntity;
+import com.odeyalo.sonata.playlists.entity.PlaylistOwnerEntity;
 import com.odeyalo.sonata.playlists.model.PlaylistType;
 import org.apache.commons.lang3.RandomStringUtils;
 
@@ -22,7 +21,7 @@ public class PlaylistEntityFaker {
         this.playlistName = playlistName;
         this.playlistDescription = playlistDescription;
         this.playlistType = playlistType;
-        this.playlistOwner = R2dbcPlaylistOwnerEntity.builder().publicId("something").displayName("odeyalo").build();
+        this.playlistOwner = PlaylistOwnerEntity.builder().publicId("something").displayName("odeyalo").build();
     }
 
     public static PlaylistEntityFaker create() {
